+++ waves/_tests/test_help_messages.py
@@ -8,12 +8,88 @@ import SCons.Environment
 from waves import scons_extensions
 
 
+# TODO: Remove this function once duplicate testing functions have been removed
+# https://re-git.lanl.gov/aea/python-projects/waves/-/issues/862
+@pytest.fixture(autouse=True)
+def reset_alias():
+    # Reset aliases before each test
+    SCons.Script.DEFAULT_TARGETS = []
+    SCons.Node.Alias.default_ans = SCons.Node.Alias.AliasNameSpace()
+
+
+# TODO: Remove entire test function when old function name `default_targets_message` is fully deprecated
+# https://re-git.lanl.gov/aea/python-projects/waves/-/issues/862
 def test_default_targets_message():
     # Raise TypeError mocking SCons < 4.6.0
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
-    with patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help:
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help,
+        patch("warnings.warn") as warning,
+    ):
         scons_extensions.default_targets_message()
+        warning.assert_called_once()
+    calls = [
+        call(ANY, "\nDefault Targets:\n", append=True, keep_local=True),
+        call(ANY, "\nDefault Targets:\n", append=True),
+    ]
+    mock_help.assert_has_calls(calls)
+
+    # No environment provided
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[None, None]) as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.default_targets_message()
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n", append=True, keep_local=True)
+
+    # Provide environment with no defaults
+    env = SCons.Environment.Environment()
+    env.Default()
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.default_targets_message(env)
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n", append=True, keep_local=True)
+
+    # Provide environment with defaults
+    env.Default("dummy.target")
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.default_targets_message(env)
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n    dummy.target\n", append=True, keep_local=True)
+
+    # Test the Method style interface
+    env.AddMethod(scons_extensions.default_targets_message, "ProjectHelp")
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        env.ProjectHelp()
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n    dummy.target\n", append=True, keep_local=True)
+
+
+def test_project_help_default_targets():
+    # Raise TypeError mocking SCons < 4.6.0
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help:
+        scons_extensions.project_help_default_targets()
     calls = [
         call(ANY, "\nDefault Targets:\n", append=True, keep_local=True),
         call(ANY, "\nDefault Targets:\n", append=True),
@@ -24,7 +100,7 @@ def test_default_targets_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[None, None]) as mock_help:
-        scons_extensions.default_targets_message()
+        scons_extensions.project_help_default_targets()
     mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n", append=True, keep_local=True)
 
     # Provide environment with no defaults
@@ -33,7 +109,7 @@ def test_default_targets_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
-        scons_extensions.default_targets_message(env)
+        scons_extensions.project_help_default_targets(env)
     mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n", append=True, keep_local=True)
 
     # Provide environment with defaults
@@ -41,11 +117,11 @@ def test_default_targets_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
-        scons_extensions.default_targets_message(env)
+        scons_extensions.project_help_default_targets(env)
     mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n    dummy.target\n", append=True, keep_local=True)
 
     # Test the Method style interface
-    env.AddMethod(scons_extensions.default_targets_message, "ProjectHelp")
+    env.AddMethod(scons_extensions.project_help_default_targets, "ProjectHelp")
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
@@ -53,12 +129,78 @@ def test_default_targets_message():
     mock_help.assert_called_once_with(ANY, "\nDefault Targets:\n    dummy.target\n", append=True, keep_local=True)
 
 
+# TODO: Remove entire test function when old function name alias_list_message is fully deprecated
+# https://re-git.lanl.gov/aea/python-projects/waves/-/issues/862
 def test_alias_list_message():
     # Raise TypeError mocking SCons < 4.6.0
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
-    with patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help:
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help,
+        patch("warnings.warn") as warning,
+    ):
         scons_extensions.alias_list_message()
+        warning.assert_called_once()
+    calls = [
+        call(ANY, "\nTarget Aliases:\n", append=True, keep_local=True),
+        call(ANY, "\nTarget Aliases:\n", append=True),
+    ]
+    mock_help.assert_has_calls(calls)
+
+    # No environment provided
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.alias_list_message()
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n", append=True, keep_local=True)
+
+    # Provide environment with no aliases
+    env = SCons.Environment.Environment()
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.alias_list_message(env)
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n", append=True, keep_local=True)
+
+    # Provide environment with alias
+    env.Alias("dummy_alias", "dummy.target")
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        scons_extensions.alias_list_message(env)
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n    dummy_alias\n", append=True, keep_local=True)
+
+    # Test the Method style interface
+    env.AddMethod(scons_extensions.alias_list_message, "ProjectHelp")
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with (
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("warnings.warn") as warning,
+    ):
+        env.ProjectHelp()
+        warning.assert_called_once()
+    mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n    dummy_alias\n", append=True, keep_local=True)
+
+
+def test_project_help_aliases():
+    # Raise TypeError mocking SCons < 4.6.0
+    # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
+    # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
+    with patch("waves.scons_extensions.SConsEnvironment.Help", side_effect=[TypeError, None]) as mock_help:
+        scons_extensions.project_help_aliases()
     calls = [
         call(ANY, "\nTarget Aliases:\n", append=True, keep_local=True),
         call(ANY, "\nTarget Aliases:\n", append=True),
@@ -69,7 +211,7 @@ def test_alias_list_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
-        scons_extensions.alias_list_message()
+        scons_extensions.project_help_aliases()
     mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n", append=True, keep_local=True)
 
     # Provide environment with no aliases
@@ -77,7 +219,7 @@ def test_alias_list_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
-        scons_extensions.alias_list_message(env)
+        scons_extensions.project_help_aliases(env)
     mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n", append=True, keep_local=True)
 
     # Provide environment with alias
@@ -85,11 +227,11 @@ def test_alias_list_message():
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
-        scons_extensions.alias_list_message(env)
+        scons_extensions.project_help_aliases(env)
     mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n    dummy_alias\n", append=True, keep_local=True)
 
     # Test the Method style interface
-    env.AddMethod(scons_extensions.alias_list_message, "ProjectHelp")
+    env.AddMethod(scons_extensions.project_help_aliases, "ProjectHelp")
     # Git commit 7a95cef7: Normally you expect something like ``patch("SCons.Script.SConscript.SConsEnvironment...")``
     # but Python <=3.10 chokes on the expected patch, so patch the WAVES module itself instead.
     with patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help:
@@ -97,6 +239,8 @@ def test_alias_list_message():
     mock_help.assert_called_once_with(ANY, "\nTarget Aliases:\n    dummy_alias\n", append=True, keep_local=True)
 
 
+# TODO: Remove entire test function when old function name project_help_message is fully deprecated
+# https://re-git.lanl.gov/aea/python-projects/waves/-/issues/862
 def test_project_help_message():
     env = SCons.Environment.Environment()
     env.AddMethod(scons_extensions.project_help_message, "ProjectHelp")
@@ -110,26 +254,83 @@ def test_project_help_message():
     method_interface_non_default_kwargs = {key: value for key, value in non_default_kwargs.items() if key != "env"}
     # Default behavior
     with (
-        patch("waves.scons_extensions.default_targets_message") as mock_targets,
-        patch("waves.scons_extensions.alias_list_message") as mock_alias,
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.SConsEnvironment.Help") as mock_help,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+        patch("warnings.warn") as warning,
     ):
         scons_extensions.project_help_message()
+        warning.assert_called_once()
         mock_targets.assert_called_once_with(**default_kwargs)
         mock_alias.assert_called_once_with(**default_kwargs)
 
     # Pass non-default kwargs
     with (
-        patch("waves.scons_extensions.default_targets_message") as mock_targets,
-        patch("waves.scons_extensions.alias_list_message") as mock_alias,
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+        patch("warnings.warn") as warning,
     ):
         scons_extensions.project_help_message(**non_default_kwargs)
+        warning.assert_called_once()
         mock_targets.assert_called_once_with(**non_default_kwargs)
         mock_alias.assert_called_once_with(**non_default_kwargs)
 
     # Test the Method style interface
     with (
-        patch("waves.scons_extensions.default_targets_message") as mock_targets,
-        patch("waves.scons_extensions.alias_list_message") as mock_alias,
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+        patch("warnings.warn") as warning,
+    ):
+        env.ProjectHelp()
+        warning.assert_called_once()
+        mock_targets.assert_called_once_with(**default_kwargs)
+        mock_alias.assert_called_once_with(**default_kwargs)
+
+    # Test the Method style interface, non-default kwargs
+    with (
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+        patch("warnings.warn") as warning,
+    ):
+        env.ProjectHelp(**method_interface_non_default_kwargs)
+        warning.assert_called_once()
+        mock_targets.assert_called_once_with(**non_default_kwargs)
+        mock_alias.assert_called_once_with(**non_default_kwargs)
+
+
+def test_project_help():
+    env = SCons.Environment.Environment()
+    env.AddMethod(scons_extensions.project_help, "ProjectHelp")
+    default_kwargs = {"env": ANY, "append": True, "keep_local": True, "target_descriptions": None}
+    non_default_kwargs = {
+        "env": env,
+        "append": False,
+        "keep_local": False,
+        "target_descriptions": {"somekey": "somevalue"},
+    }
+    method_interface_non_default_kwargs = {key: value for key, value in non_default_kwargs.items() if key != "env"}
+    # Default behavior
+    with (
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+    ):
+        scons_extensions.project_help()
+        mock_targets.assert_called_once_with(**default_kwargs)
+        mock_alias.assert_called_once_with(**default_kwargs)
+
+    # Pass non-default kwargs
+    with (
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
+    ):
+        scons_extensions.project_help(**non_default_kwargs)
+        mock_targets.assert_called_once_with(**non_default_kwargs)
+        mock_alias.assert_called_once_with(**non_default_kwargs)
+
+    # Test the Method style interface
+    with (
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
     ):
         env.ProjectHelp()
         mock_targets.assert_called_once_with(**default_kwargs)
@@ -137,8 +338,8 @@ def test_project_help_message():
 
     # Test the Method style interface, non-default kwargs
     with (
-        patch("waves.scons_extensions.default_targets_message") as mock_targets,
-        patch("waves.scons_extensions.alias_list_message") as mock_alias,
+        patch("waves.scons_extensions.project_help_default_targets") as mock_targets,
+        patch("waves.scons_extensions.project_help_aliases") as mock_alias,
     ):
         env.ProjectHelp(**method_interface_non_default_kwargs)
         mock_targets.assert_called_once_with(**non_default_kwargs)