+++ waves/_tests/test_scons_extensions.py
@@ -5,7 +5,7 @@ import copy
 import pathlib
 from contextlib import nullcontext as does_not_raise
 import unittest
-from unittest.mock import patch, call
+from unittest.mock import patch, call, Mock
 import subprocess
 
 import pytest
@@ -32,6 +32,35 @@ from waves._tests.common import platform_check
 fs = SCons.Node.FS.FS()
 
 testing_windows, root_fs, testing_macos = platform_check()
+
+
+mock_decodable = Mock()
+mock_decodable.__str__ = lambda self: "mock_node"
+mock_decodable.get_executor.return_value.get_contents.return_value = b"action signature string"
+mock_not_decodable = Mock()
+mock_not_decodable.__str__ = lambda self: "mock_node"
+mock_not_decodable.get_executor.return_value.get_contents.return_value = b"\x81action signature string"
+test_print_action_signature_string_cases = {
+    "decode-able": (mock_decodable, "action signature string"),
+    "not decode-able": (mock_not_decodable, b"\x81action signature string"),
+}
+
+
+@pytest.mark.parametrize(
+    "mock_node, action_signature_string",
+    test_print_action_signature_string_cases.values(),
+    ids=test_print_action_signature_string_cases.keys(),
+)
+def test_print_action_signature_string(mock_node, action_signature_string):
+    s = "s"
+    source = []
+    env = SCons.Environment.Environment()
+    with patch("builtins.print") as mock_print:
+        target = [mock_node]
+        scons_extensions.print_action_signature_string(s, target, source, env)
+        mock_print.assert_called_once_with(
+            f"Building {mock_node} with action signature string:\n  {action_signature_string}\n{s}",
+        )
 
 
 check_program = {
@@ -816,6 +845,7 @@ construct_action_list = {
     "list2": (["thing1", "thing2"], prefix, "", [f"{prefix} thing1", f"{prefix} thing2"]),
     "tuple": (("thing1",), prefix, "", [f"{prefix} thing1"]),
     "str": ("thing1", prefix, "", [f"{prefix} thing1"]),
+    "pathlib.Path": (pathlib.Path("thing1"), prefix, "", [f"{prefix} thing1"]),
     "list1 suffix": (["thing1"], prefix, suffix, [f"{prefix} thing1 {suffix}"]),
     "list2 suffix": (["thing1", "thing2"], prefix, suffix, [f"{prefix} thing1 {suffix}", f"{prefix} thing2 {suffix}"]),
     "tuple suffix": (("thing1",), prefix, suffix, [f"{prefix} thing1 {suffix}"]),
@@ -1144,6 +1174,14 @@ abaqus_pseudobuilder_input = {
         " job=job double=both $(cpus=1$)",
         {},
     ),
+    "job, subdirectory": (
+        {},
+        {"job": f"subdir{os.path.sep}job"},
+        [f"subdir{os.path.sep}job.inp"],
+        [f"subdir{os.path.sep}job{ext}" for ext in _abaqus_standard_extensions],
+        " job=job double=both $(cpus=1$)",
+        {},
+    ),
     "override cpus": (
         {"override_cpus": 2},
         {"job": "job", "cpus": 3},
@@ -1236,7 +1274,7 @@ abaqus_pseudobuilder_input = {
     abaqus_pseudobuilder_input.values(),
     ids=abaqus_pseudobuilder_input.keys(),
 )
-def test_abaqus_pseudobuilder(class_kwargs, call_kwargs, sources, targets, options, builder_kwargs):
+def test_abaqus_pseudo_builder(class_kwargs, call_kwargs, sources, targets, options, builder_kwargs):
     # Mock AbaqusSolver builder and env
     mock_builder = unittest.mock.Mock()
     mock_env = unittest.mock.Mock()
@@ -1427,7 +1465,7 @@ builder_factory_tests.update(
             "environment": "",
             "action_prefix": _cd_action_prefix,
             "program": "abaqus",
-            "program_required": "-interactive -ask_delete no -input ${SOURCE.filebase}",
+            "program_required": "-interactive -ask_delete no -job ${job} -input ${SOURCE.filebase}",
             "program_options": "",
             "subcommand": "",
             "subcommand_required": "",
@@ -2286,7 +2324,7 @@ parameter_study_write_cases = {
         {},
         [],
         pytest.raises(RuntimeError),
-    )
+    ),
 }
 
 
@@ -2359,11 +2397,12 @@ waves_environment_methods = {
     "AddCubit": ("AddCubit", "add_cubit"),
     "AddCubitPython": ("AddCubitPython", "add_cubit_python"),
     "CopySubstfile": ("CopySubstfile", "copy_substfile"),
-    "ProjectHelp": ("ProjectHelp", "project_help_message"),
+    "ProjectHelp": ("ProjectHelp", "project_help"),
     "ProjectAlias": ("ProjectAlias", "project_alias"),
     "SubstitutionSyntax": ("SubstitutionSyntax", "substitution_syntax"),
     "ParameterStudyTask": ("ParameterStudyTask", "parameter_study_task"),
     "ParameterStudySConscript": ("ParameterStudySConscript", "parameter_study_sconscript"),
+    "ParameterStudyWrite": ("ParameterStudyWrite", "parameter_study_write"),
 }
 
 
@@ -2437,3 +2476,13 @@ def test_waves_environment_builders(builder, factory, factory_kwargs):
         attribute(target, source, *args, **kwargs)
         mock_factory.assert_called_once_with(**factory_kwargs)
         mock_builder.assert_called_once_with(env, *args, target=target, source=source, **kwargs)
+
+
+def test_waves_environment_abaqus_pseudo_builder():
+    env = scons_extensions.WAVESEnvironment()
+
+    args = ["arg1"]
+    kwargs = {"kwarg1": "value1"}
+    with patch("waves.scons_extensions.AbaqusPseudoBuilder.__call__") as mock_call:
+        env.AbaqusPseudoBuilder("job", *args, **kwargs)
+        mock_call.assert_called_once_with(env, "job", *args, **kwargs)